requirements.txt
pymongo==4.6.1
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
Pillow==10.1.0
reportlab==4.0.8
seaborn==0.13.1
openpyxl==3.1.2
contourpy>=1.0.1
cycler>=0.10
fonttools>=4.22.0
kiwisolver>=1.0.1
packaging>=20.0
pyparsing>=2.3.1
python-dateutil>=2.7 